Corrected a false warning for the Stochastic loss function.
Sometimes the probability for the Bremsstrahlung crossection is just
zero because of the cutoff paremeters. The function now checks the
integral limits for all components. The probabilities are only zero if
for all components the upper integration limit is equal to the lower
integration limit.
If this is not the case then the probabilities weren't calculated or
calculated wrong.