semisolve the problem with total_rate=0
if the particle energy is to low, the probability for an interaction vanishes. (due to the cuts)
in PROPOSAL version 1 then the initial energy was set to the final energy and the particle was propagated again.
in version 2 this was not adopt; now the continue in the while loop is implemented.

BUT: I don't get, why the probability vanishes in Ionization; this should be aprrox. constant over the whole energy range?
so this has to be fixed in the future