feat: add 2 methods in MuonPropagator
For complete comaptibility to icesim, the 2 methods:
GetStochasticRate and GetTotalStochasticRate
which were till now uncommented in the new versions are now available again.
The parameter I3Particle::ParticleType type is not needed in the new code structure.
It can be deleted in the future and is just left inside because of backward compatibility.