feat: I3: Finalize propagation with stoch Eloss
The I3PropagatorServicePROPOSAL has two new arguments
"final_stochastic_loss" and "distance_to_propagte".
The "final_stochastic_loss" is an I3Particle::ParticleType.
If this argument is given and is not I3Particle::unknown,
the remaining energy, after propagating the
particle the distance "distance_to_propagte", will be
stored in an I3Particle of the given type
"final_stochastic_loss". This particle is appended to
the secondaries vector.