feat: add ability to sample uniform phase space
This feature effects the many body decay for
especially tau leptons. In this decay method the
momenta weren't sampled uniformly in the phase space because
of the intrinsic phase space density in this algorithm.
This commit add the ability to sample these momenta uniformly by
rejection sampling.
This ability can be set in the decay table of the particle definitions
or by a key "uniform" in the config file.