feat: add a new leptonic decay channel
There was only one leptonic decay channel available for
decays like "tau -> mu nu nu" or "tau -> e nu nu".
This channel is valid if the mass of the produced massive
lepton is low compared to the parent lepton.
But this is not the case for "tau -> mu".
So a new, more accurate, channel is added for this case.