replacing lots of critical boost functions
replace the json readout
the boost property tree was causing problems on sl6 and boost4.84,
therfore it was replaced by the single_header include json reader by nlohman

also the boost erf_inv (calculation of the inverse error function)
was causing problems with using valgrind
so it was replaced by a commonly used method

and some c++11 features are used,
so the std::round is replacing the self written round
or the std::min for a list of 3 parameters

The Sector::Definition gets an assignment operator.

The config file was slightly changed to a more intuitive usage.

My apologies for the large commit