feature: pybind MakeSochasticLoss
to include MakeStochasticLoss and CalculateEnergyTillStochastic in Sector UnitTest
the MakeStochasticLoss nedded to be pybinded.
I think it's also more transparent, if the MakeStochasticLoss get the particle energy explicit as an argument.
The TestFiles are not changed yet, as the Tests fail anyway.
But hte python script to generate the script is also adaptedfeature: pybind MakeSochasticLoss

to include MakeStochasticLoss and CalculateEnergyTillStochastic in Sector UnitTest
the MakeStochasticLoss nedded to be pybinded.
I think it's also more transparent, if the MakeStochasticLoss get the particle energy explicit as an argument.
The TestFiles are not changed yet, as the Tests fail anyway.
But hte python script to generate the script is also adapted.